openjpa-lib/src/test/java/org/apache/openjpa/lib/log/JULLogFactoryTest.java
/*
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */
package org.apache.openjpa.lib.log;

import org.junit.Test;

import java.util.ArrayList;
import java.util.Collection;
import java.util.logging.Handler;
import java.util.logging.LogRecord;
import java.util.logging.Logger;

import static org.junit.Assert.assertEquals;

public class JULLogFactoryTest {
    @Test
    public void log() {
        final JULLogFactory factory = new JULLogFactory();
        final String logger = getClass().getName() + "__log";
        final Logger jul = Logger.getLogger(logger);
        final Collection<LogRecord> records = new ArrayList<>();
        final Handler handler = new Handler() {
            @Override
            public void publish(final LogRecord record) {
                records.add(record);
            }

            @Override
            public void flush() {
                // no-op
            }

            @Override
            public void close() throws SecurityException {
                flush();
            }
        };
        jul.addHandler(handler);
        final Log log = factory.getLog(logger);
        log.info(">test<");
        jul.removeHandler(handler);
        assertEquals(1, records.size());
        assertEquals(">test<", records.iterator().next().getMessage());
    }
}
